sprint management

+++ app/Http/Controllers/UserController.php
@@ -9,6 +9,7 @@ use App\Http\Requests\User\UpdateUserRequest;
 use App\Models\User;
 use Exception;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Hash;
 use Illuminate\Support\Facades\Log;
 
@@ -17,9 +18,9 @@ class UserController extends Controller
     public function store(NewUserRequest $request)
     {
         try {
-          $user=  User::create($request->all());
-           $user->assignRole($request->role);
-           
+            $user =  User::create($request->all());
+            $user->assignRole($request->role);
+
             return response()->json(['status' => ResponseStatus::SUCCESS, 'message' => 'User Created'], 200);
         } catch (Exception $e) {
             return response()->json(['status' => ResponseStatus::ERROR, 'message' => $e->getMessage()], 500);
@@ -35,14 +36,13 @@ class UserController extends Controller
                     'user' => $user,
                     'message' => 'User Found',
                 ], 200);
-            }else{
+            } else {
                 return response()->json([
                     'status' => ResponseStatus::SUCCESS,
                     'user' => null,
                     'message' => 'User Not Found',
                 ], 200);
             }
-
         } catch (Exception $e) {
             return response()->json([
                 'status' => ResponseStatus::ERROR,
@@ -54,13 +54,13 @@ class UserController extends Controller
     public function edit(User $user, UpdateUserRequest $request)
     {
         try {
-                    
-   
-            $user->update($request->all());
 
-            if($request->role!=='')$user->syncRoles($request->role);
-            
-            return response()->json(['status' => ResponseStatus::SUCCESS, 'message' => 'User Edited'], 200);
+            if ($user->hasRole(Role::ADMIN)) {
+                $user->update($request->all());
+                if ($request->role !== '') $user->syncRoles($request->role);
+            } else {
+                return response()->json(['status' => ResponseStatus::ERROR, 'message' => 'Invalid Access'], 401);
+            }
         } catch (Exception $e) {
             return response()->json(['status' => ResponseStatus::ERROR, 'message' => $e->getMessage()], 500);
         }
@@ -68,7 +68,18 @@ class UserController extends Controller
     public function destroy(User $user)
     {
         try {
-            $user->delete();
+
+            if ($user->hasRole(Role::ADMIN)) {
+                if (Auth::user()->id !== $user->id) {
+                    $user->delete();
+                    return response()->json(['status' => ResponseStatus::SUCCESS, 'message' => 'User Deleted Successfully'], 200);
+                } else {
+                    return response()->json(['status' => ResponseStatus::ERROR, 'message' => 'Cannot delete the logged in user'], 500);
+                }
+            } else {
+                return response()->json(['status' => ResponseStatus::ERROR, 'message' => 'Invalid Access'], 401);
+            }
+
             return response()->json(['status' => ResponseStatus::SUCCESS, 'message' => 'User Destroyed'], 200);
         } catch (Exception $e) {
             return response()->json(['status' => ResponseStatus::ERROR, 'message' => $e->getMessage()], 500);
@@ -77,21 +88,20 @@ class UserController extends Controller
     public function all(Request $request)
     {
         try {
-            $users =User::with('roles')->get();
+            $users = User::with('roles')->get();
             if ($users) {
                 return response()->json([
                     'status' => ResponseStatus::SUCCESS,
                     'users' => $users,
                     'message' => 'Users Found',
                 ], 200);
-            }else{
+            } else {
                 return response()->json([
                     'status' => ResponseStatus::SUCCESS,
                     'users' => null,
                     'message' => 'Users Not Found',
                 ], 200);
             }
-
         } catch (Exception $e) {
             return response()->json([
                 'status' => ResponseStatus::ERROR,
